Showing result done

+++ app/src/main/java/com/pmondal/weather/models/ForecastWeatherModel.kt
@@ -27,9 +27,9 @@ data class ForecastModel(
         @SerializedName("population")
         val population: Int,
         @SerializedName("sunrise")
-        val sunrise: Int,
+        val sunrise: Long,
         @SerializedName("sunset")
-        val sunset: Int,
+        val sunset: Long,
         @SerializedName("timezone")
         val timezone: Int
     ) {
@@ -45,7 +45,7 @@ data class ForecastModel(
         @SerializedName("clouds")
         val clouds: Clouds,
         @SerializedName("dt")
-        val dt: Int,
+        val dt: Long,
         @SerializedName("dt_txt")
         val dtTxt: String,
         @SerializedName("main")
